Uodate
Updated the dockfile to install the zip-extension, which is required for
using the class "ZipArchive".
Tried to generate a zipfile, which was a success. We have extension..!
Noticed that the name of the pdf-files was only the email of the user in
the zip-archive, without any valid extension for a pdf-file. Shall look
into if that is something we are to fix or if it is by design.