Update
Updated the docker-file to include locales and set the correct locale.
Updated index.php to setlocale() for sv_SE.
config.php got a change of locale to sv_SE, so we can use the
default-value if we wish.
The rest of the files has just been touched and contains no direct
changes.